Main group 20 client product screen

+++ back-end/src/database/models/Product.js
@@ -2,7 +2,7 @@ const ProductSchema = (sequelize, DataTypes) => {
   const ProductTable = sequelize.define('Product', {
     id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
     name: DataTypes.STRING,
-    price: DataTypes.FLOAT,
+    price: DataTypes.DECIMAL(4,2),
     urlImage: DataTypes.STRING,
   }, {
     tableName: 'products',
